daemon/cryptopkt: don't use struct pkt.
It's unnecessary; just keep a separate length and data pointer.

One minor change is that we *always* place into peer->inpkt; we were
always going to, though it's a slight layering violation to assume it.
Otherwise we'd have to find somewhere to stash the length.

It also fixes the data input count, so we can send more than 1 packet :)

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>