daemon/cryptopkt: two stage handshake.
Both Anthony Towns and matsjj (of thunder.network fame) disliked the
information revelation of sending our pubkey in the clear.  This
implements Mats's idea of sending session key first, then proving
it's us in a second round.

This means more work for the peer in the DoS case.  It also adds some
optional public data (currently we send none).

Suggested-by: Mats Jerratsch <matsjj@gmail.com>
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>